LeadMe_back/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.30.0
numpy==1.24.3
pandas==2.0.1
python-multipart==0.0.6
boto3==1.26.133
pytest==7.3.1
httpx==0.24.0
pillow==9.5.0